Short endless runner game to practice Audio Programming.
In this game, created all assets and audio.
The Background Music is split into 3 loops. Each loop is played from start, and a layer is removed every time a life is lost.
The Music also increases in pitch and tempo over time.

This was my first c# and Unity project.